misc code cleanup
